backend/requirements.txt
appdirs==1.4.4
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cssselect2==0.8.0
Flask==3.1.2
fonttools==4.60.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
JayDeBeApi==1.2.3
Jinja2==3.1.6
jpype1==1.6.0
MarkupSafe==3.0.2
packaging==25.0
pdfkit==1.0.0
pillow==11.3.0
pycparser==2.23
pydyf==0.11.0
pyee==11.1.1
pypdf==6.1.1
PyPDF2==3.0.1
pyphen==0.17.2
pyppeteer==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
reportlab==4.4.4
requests==2.32.5
six==1.17.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==1.26.20
weasyprint==66.0
webencodings==0.5.1
websockets==10.4
Werkzeug==3.1.3
zipp==3.23.0
zopfli==0.2.3.post1
